+++ alinkjs/packages/assetlink-plugin-api/src/components/EntitySearch.vue
@@ -1,25 +1,28 @@
 <script>
 /**
- * Provides a UI for searching for and choosing assets using
+ * Provides a UI to search for and choose entities using
  * plugin-supplied search methods.
  *
  * ### Usage
  *
  * ```js
- * <asset-selector
+ * <entity-search
  *   title="Find Asset"
  *   &commat;submit="(assets) => onAssetSelected(assets)"
- * ></asset-selector>
+ * ></entity-search>
  * ```
  *
  * @category components
  * @vue-prop {String} title - the title to show at the top of the selector
+ * @vue-prop {String} entityType - the entity type to search
+ * @vue-prop {String[]} [entityBundles=null] - the entity bundles to search from - or all if null
  * @vue-prop {String} [searchMethod=text-search] - the initially selected search method
- * @vue-prop {Boolean} [selectMultiple=false] - allow selecting multiple assets
- * @vue-prop {String} [confirmLabel=Choose] - the text of the button which confirms the current asset selection
- * @vue-prop {Object} [additionalFilters=[]] - Additional [Orbit.js filters]{@link https://orbitjs.com/docs/querying-data#attribute-filtering} to apply to the searches
+ * @vue-prop {Boolean} [multiple=false] - allow selecting multiple entities
+ * @vue-prop {String} [confirmLabel=Choose] - the text of the button which confirms the current entity selection
+ * @vue-prop {String} [noResultsLabel=No results found] - the text shown when no results are found for a given search
+ * @vue-prop {Object[]} [additionalFilters=[]] - Additional [Orbit.js filters]{@link https://orbitjs.com/docs/querying-data#attribute-filtering} to apply to the searches
  * @vue-event {String} searchMethodChanged - Emit currently selected search method
- * @vue-event {Asset[]} submit - Emit selected asset(s)
+ * @vue-event {Entity[]} submit - Emit selected entities
  */
 export default {};
 </script>
@@ -31,9 +34,12 @@ import RacingLocalRemoteAsyncIterator from "../RacingLocalRemoteAsyncIterator";
 
 const props = defineProps({
   title: { type: String, required: true },
+  entityType: { type: String, required: true },
+  entityBundles: { type: Array, default: null },
   searchMethod: { type: String, default: "text-search" },
-  selectMultiple: { type: Boolean, default: false },
+  multiple: { type: Boolean, default: false },
   confirmLabel: { type: String, default: "Choose" },
+  noResultsLabel: { type: String, default: "No results found" },
   additionalFilters: { type: Array, default: () => [] },
 });
 
@@ -42,7 +48,7 @@ const route = useRoute();
 const assetLink = inject("assetLink");
 
 const searchMethodTileDefs = computed(() =>
-  assetLink.getSlots({ type: "asset-search-method", route })
+  assetLink.getSlots({ type: `${props.entityType}-search-method`, route })
 );
 
 const currentSearchMethod = ref(props.searchMethod);
@@ -62,33 +68,33 @@ const searchResultEntries = ref([]);
 const selectedKey = ref(null);
 const tickedKeys = ref([]);
 
-const selectedAssets = computed(() => {
-  const searchResultAssets = searchResultEntries.value.map(
-    (entry) => entry.asset
+const selectedEntities = computed(() => {
+  const searchResultEntities = searchResultEntries.value.map(
+    (entry) => entry.entity
   );
   if (selectedKey.value) {
-    return searchResultAssets.filter((asset) => asset.id === selectedKey.value);
+    return searchResultEntities.filter(
+      (entity) => entity.id === selectedKey.value
+    );
   }
-  return searchResultAssets.filter((asset) =>
-    tickedKeys.value.includes(asset.id)
+  return searchResultEntities.filter((entity) =>
+    tickedKeys.value.includes(entity.id)
   );
 });
 
-const hasAssetSelection = computed(() => {
-  return selectedAssets.value.length > 0;
+const hasEntitySelection = computed(() => {
+  return selectedEntities.value.length > 0;
 });
-
-const isSearchingAssets = ref(false);
 
 const searchRequest = ref(undefined);
 
 const nodes = computed(() => {
   return searchResultEntries.value.map((entry) => {
     return {
-      id: entry.asset.id,
-      asset: entry.asset,
+      id: entry.entity.id,
+      entity: entry.entity,
       weightText: entry.weightText,
-      tickable: props.selectMultiple ? true : false,
+      tickable: props.multiple ? true : false,
     };
   });
 });
@@ -96,8 +102,10 @@ const nodes = computed(() => {
 // TODO: Add a "show more" button at the bottom of the search to increase this
 const maxDesiredSearchEntries = 10;
 
-const searchAssets = async function searchAssets() {
+const searchEntities = async function searchEntities() {
   const currSearchReq = {
+    entityType: props.entityType,
+    entityBundles: props.entityBundles,
     ...searchRequest.value,
   };
 
@@ -106,14 +114,15 @@ const searchAssets = async function searchAssets() {
     ...props.additionalFilters,
   ];
 
-  isSearchingAssets.value = true;
-
-  let assetSearchResultCursor = assetLink.searchAssets(currSearchReq, "local");
+  let entitySearchResultCursor = assetLink.searchEntities(
+    currSearchReq,
+    "local"
+  );
 
   if (assetLink.connectionStatus.isOnline) {
-    assetSearchResultCursor = new RacingLocalRemoteAsyncIterator(
-      assetSearchResultCursor,
-      assetLink.searchAssets(currSearchReq, "remote")
+    entitySearchResultCursor = new RacingLocalRemoteAsyncIterator(
+      entitySearchResultCursor,
+      assetLink.searchEntities(currSearchReq, "remote")
     );
   }
 
@@ -123,14 +132,14 @@ const searchAssets = async function searchAssets() {
 
   searchResultEntries.value = [];
 
-  const alreadyFoundAssetIds = new Set();
+  const alreadyFoundEntityIds = new Set();
 
   let searchIterItem = {};
   while (
     searchResultEntries.value.length < maxDesiredSearchEntries &&
     !searchIterItem.done
   ) {
-    searchIterItem = await assetSearchResultCursor.next();
+    searchIterItem = await entitySearchResultCursor.next();
 
     if (currSearchReq.id !== searchRequest.value.id) {
       return;
@@ -138,23 +147,21 @@ const searchAssets = async function searchAssets() {
 
     if (
       searchIterItem.value &&
-      !alreadyFoundAssetIds.has(searchIterItem.value.asset.id)
+      !alreadyFoundEntityIds.has(searchIterItem.value.entity.id)
     ) {
-      alreadyFoundAssetIds.add(searchIterItem.value.asset.id);
+      alreadyFoundEntityIds.add(searchIterItem.value.entity.id);
       searchResultEntries.value.push(searchIterItem.value);
     }
   }
-
-  isSearchingAssets.value = false;
 };
 
 watch(searchRequest, (val) => {
-  searchAssets();
+  searchEntities();
   emit("update:searchRequest", val);
 });
 
 const onAccept = () => {
-  emit("submit", selectedAssets.value);
+  emit("submit", selectedEntities.value);
 };
 
 const onCancel = () => {
@@ -191,14 +198,14 @@ const onCancel = () => {
       <q-tree
         node-key="id"
         :nodes="nodes"
-        :tick-strategy="props.selectMultiple ? 'leaf' : 'none'"
+        :tick-strategy="props.multiple ? 'leaf' : 'none'"
         v-model:ticked="tickedKeys"
         v-model:selected="selectedKey"
-        no-nodes-label="No assets found"
+        :no-nodes-label="props.noResultsLabel"
         class="q-mx-lg q-mb-md"
       >
         <template v-slot:default-header="prop">
-          <entity-name :entity="prop.node.asset"></entity-name>
+          <entity-name :entity="prop.node.entity"></entity-name>
         </template>
 
         <template v-slot:default-body="prop">
@@ -217,7 +224,7 @@ const onCancel = () => {
         color="primary"
         :label="props.confirmLabel"
         @click="onAccept()"
-        :disabled="!hasAssetSelection"
+        :disabled="!hasEntitySelection"
       />
     </div>
   </div>